i found the solution:
http://serverfault.com/questions/573025/taskset-not-working-over-a-range-of-cores-in-isolcpus

This behavior seems to be an artifact of the default kernel scheduler
algorithm (SCHED_OTHER for this distro/kernel). Changing the process
to a different algorithm eliminates the problem, isolcpus are
adequately utilized across the processes/threads.

$ sudo chrt -r 1 [command]

(If you want to run as non-root you can use the setcap utility to
enable CAP_SYS_NICE on the binary file related to the command)